Literacy Rate in Tamilnadu
This dashboard empowers mission driven organizations to harness the power of data visualization for social change. The data set talks about the literacy rates in Tamilnadu. It is distributed according to the cities and their ranks. It contains the population, as per the gender.

This dashboard gives us the insight about the female population when compared to the total, and it addresses the question of where the state is on literacy wise. It shows us the status of each city and it also shows the scholarship/Beneficiaries each state is getting. The data has different institutions available and the number of teachers available.

General Requirements
The dashboard should look like an article predicting the basic issues. Use appropriate tiled and floating sheets wherever possible.

Business Requirements
Title: Literacy Rate in Tamilnadu